refactor: move store to heap
