+++ src/observatory/src/lib.rs
@@ -7,7 +7,6 @@ mod upgrade;
 
 use crate::guards::caller_is_admin_controller;
 use crate::memory::STATE;
-use crate::store::{delete_controllers, set_controllers as set_controllers_store};
 use crate::types::state::{HeapState, State};
 use crate::upgrade::types::upgrade::UpgradeStableState;
 use ic_cdk::caller;
@@ -15,6 +14,7 @@ use ic_cdk::storage::{stable_restore, stable_save};
 use ic_cdk_macros::{export_candid, init, post_upgrade, pre_upgrade, query, update};
 use junobuild_shared::controllers::init_controllers;
 use junobuild_shared::types::interface::{DeleteControllersArgs, SetControllersArgs};
+use store::heap::{delete_controllers, set_controllers as set_controllers_store};
 
 #[init]
 fn init() {
